+++ Api/src/member-module-level/entities/member-module-level.entity.ts
@@ -16,11 +16,11 @@ import { Member } from '../../member/entities/member.entity';
 export class MemberModuleLevel {
   // Primary key information
   @PrimaryColumn({ type: 'bigint' })
-  @Field(() => Int)
+  @Field(() => Float)
   guild_id: number;
 
   @PrimaryColumn({ type: 'bigint' })
-  @Field(() => Int)
+  @Field(() => Float)
   user_id: number;
 
   // Relations
